requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
PyJWT==1.7.1
PyMySQL==0.9.3
requests==2.22.0
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
